+++ server/sonar-server/src/test/java/org/sonar/server/setting/ws/SetActionTest.java
@@ -355,13 +355,13 @@ public class SetActionTest {
 
   @Test
   public void user_setting_is_not_updated() {
-    propertyDb.insertProperty(newGlobalPropertyDto("my.key", "my user value").setUserId(42L));
+    propertyDb.insertProperty(newGlobalPropertyDto("my.key", "my user value").setUserId(42));
     propertyDb.insertProperty(newGlobalPropertyDto("my.key", "my global value"));
 
     callForGlobalSetting("my.key", "my new global value");
 
     assertGlobalSetting("my.key", "my new global value");
-    assertUserSetting("my.key", "my user value", 42L);
+    assertUserSetting("my.key", "my user value", 42);
   }
 
   @Test
@@ -826,8 +826,8 @@ public class SetActionTest {
       .containsExactly(key, value, null);
   }
 
-  private void assertUserSetting(String key, String value, long userId) {
-    List<PropertyDto> result = dbClient.propertiesDao().selectByQuery(PropertyQuery.builder().setKey(key).setUserId((int) userId).build(), dbSession);
+  private void assertUserSetting(String key, String value, int userId) {
+    List<PropertyDto> result = dbClient.propertiesDao().selectByQuery(PropertyQuery.builder().setKey(key).setUserId(userId).build(), dbSession);
 
     assertThat(result).hasSize(1)
       .extracting(PropertyDto::getKey, PropertyDto::getValue, PropertyDto::getUserId)
