+++ sonar-db/src/main/java/org/sonar/db/user/GroupMembershipDao.java
@@ -34,12 +34,12 @@ import static org.sonar.db.DatabaseUtils.executeLargeInputs;
 
 public class GroupMembershipDao implements Dao {
 
-  public List<GroupMembershipDto> selectGroups(DbSession session, GroupMembershipQuery query, Long userId, int offset, int limit) {
+  public List<GroupMembershipDto> selectGroups(DbSession session, GroupMembershipQuery query, Integer userId, int offset, int limit) {
     Map<String, Object> params = ImmutableMap.of("query", query, "userId", userId);
     return mapper(session).selectGroups(params, new RowBounds(offset, limit));
   }
 
-  public int countGroups(DbSession session, GroupMembershipQuery query, Long userId) {
+  public int countGroups(DbSession session, GroupMembershipQuery query, Integer userId) {
     Map<String, Object> params = ImmutableMap.of("query", query, "userId", userId);
     return mapper(session).countGroups(params);
   }
@@ -54,7 +54,7 @@ public class GroupMembershipDao implements Dao {
     return mapper(session).countMembers(params);
   }
 
-  public Map<String, Integer> countUsersByGroups(DbSession session, Collection<Long> groupIds) {
+  public Map<String, Integer> countUsersByGroups(DbSession session, Collection<Integer> groupIds) {
     Map<String, Integer> result = Maps.newHashMap();
     executeLargeInputs(
       groupIds,
@@ -69,7 +69,7 @@ public class GroupMembershipDao implements Dao {
     return result;
   }
 
-  public List<Long> selectGroupIdsByUserId(DbSession dbSession, long userId) {
+  public List<Integer> selectGroupIdsByUserId(DbSession dbSession, int userId) {
     return mapper(dbSession).selectGroupIdsByUserId(userId);
   }
